Prepare to release 0.3 (#26)
* 1) Recompile.
2) Update WizardsForVaading sub module.

* Update README.md

* Fix bug with duplicate history.

* WIP for trace matrix

* WIP for trace matrix

* WIP for trace matrix

* WIP for trace matrix

* Re-add sub module.

* Re-add sub module.

* Changes in Travis script.

* Changes in Travis script.

* Changes in Travis script.

* Try to fix broken build.

* CHange dependency so it compiles on brand new repository.

* Change to correct branch.

* Right version.

* A fix?

* Fix bug associating requirement version to a test case.

* Progress on the Trace Matrix.

* Add coverage column.

* 1) Fix issue with root node menu.
2) Fix issue creating a new project causing NPE.

* Fix issue where executions including Test Cases already in previous ones wouldn't show up.

* Move the requirement version assignment from execution of the step to the creation of the execution.

* Add requirement version to created executions.

* Handle possible NPE.

* Add icon to message box.

* Add coverity stuff to Travis script.

* Add message.

* Update parent/children relationship of requirements.

* Allow to display parent and child requirements.

* Fix formatting on the string representation.

* Enhance demo project to include product spec requirements and links to SW requirements.

* Enhancements to display of parent and children requirements.

* Better demo project. Add requirements links to all requirements for a better populated trace matrix.

* More detailed Trace Matrix.

* 1) Refactor method to display and select requirements from a list.
2) Add feature to add children to a requirement from GUI.

* Required for coverity commands.

* Fix bug where local changes were being lost.

* Fixing test failure. The local object was not updated with import changes so following calls would import all all over again.

* More test fixes.

* More test fixes.

* Add history to steps and link one to the execution step.

* Add step history to steps.
Task #22 - Provide UI features for Execution step management

* Add menus to create and delete executions.
Task #22 - Provide UI features for Execution step management

* Try to fix issues on Travis.

* Fix errors during test.

* Enable cobertura.

* Expose method as public and static.

* Add read requirement permission to quality role.

* Remove redundant class AuditedObject.

* Make sure to change reason.

* Add missing overrides.

* Add test for fixes in versioning system.

* Fix some issues related to tying user with changes.
Task #23 - Implement change history reasons

* Add some missing fields on the update method.
Task #25 - history seems to disappear

* Update dependencies.

* Update dependencies.

* Replicate issue #25
Task #25 - history seems to disappear

* Add checks to verify the History entries are present or not.
Task #25 - history seems to disappear

* Refactoring.

* Might be versionable and not have history (i.e. no fields marked as auditable). In that case set default values.

* Fix inaccurate comment.

* Set entity manager to null when closing.

* Finally fix bug
Task #25 - history seems to disappear

* Update to Vaadin 7.7.9.

* Update database.

* Release actions.

* Fix some code review observations.

+++ VM-Core/src/main/java/com/validation/manager/core/db/controller/VmSettingJpaController.java
@@ -12,6 +12,7 @@ import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
 import com.validation.manager.core.db.History;
 import com.validation.manager.core.db.VmSetting;
+import com.validation.manager.core.db.controller.exceptions.IllegalOrphanException;
 import com.validation.manager.core.db.controller.exceptions.NonexistentEntityException;
 import java.util.ArrayList;
 import java.util.List;
@@ -49,8 +50,13 @@ public class VmSettingJpaController implements Serializable {
             vmSetting.setHistoryList(attachedHistoryList);
             em.persist(vmSetting);
             for (History historyListHistory : vmSetting.getHistoryList()) {
-                historyListHistory.getVmSettingList().add(vmSetting);
+                VmSetting oldVmSettingIdOfHistoryListHistory = historyListHistory.getVmSettingId();
+                historyListHistory.setVmSettingId(vmSetting);
                 historyListHistory = em.merge(historyListHistory);
+                if (oldVmSettingIdOfHistoryListHistory != null) {
+                    oldVmSettingIdOfHistoryListHistory.getHistoryList().remove(historyListHistory);
+                    oldVmSettingIdOfHistoryListHistory = em.merge(oldVmSettingIdOfHistoryListHistory);
+                }
             }
             em.getTransaction().commit();
         }
@@ -61,7 +67,7 @@ public class VmSettingJpaController implements Serializable {
         }
     }
 
-    public void edit(VmSetting vmSetting) throws NonexistentEntityException, Exception {
+    public void edit(VmSetting vmSetting) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -69,6 +75,18 @@ public class VmSettingJpaController implements Serializable {
             VmSetting persistentVmSetting = em.find(VmSetting.class, vmSetting.getId());
             List<History> historyListOld = persistentVmSetting.getHistoryList();
             List<History> historyListNew = vmSetting.getHistoryList();
+            List<String> illegalOrphanMessages = null;
+            for (History historyListOldHistory : historyListOld) {
+                if (!historyListNew.contains(historyListOldHistory)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain History " + historyListOldHistory + " since its vmSettingId field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
             List<History> attachedHistoryListNew = new ArrayList<History>();
             for (History historyListNewHistoryToAttach : historyListNew) {
                 historyListNewHistoryToAttach = em.getReference(historyListNewHistoryToAttach.getClass(), historyListNewHistoryToAttach.getId());
@@ -77,16 +95,15 @@ public class VmSettingJpaController implements Serializable {
             historyListNew = attachedHistoryListNew;
             vmSetting.setHistoryList(historyListNew);
             vmSetting = em.merge(vmSetting);
-            for (History historyListOldHistory : historyListOld) {
-                if (!historyListNew.contains(historyListOldHistory)) {
-                    historyListOldHistory.getVmSettingList().remove(vmSetting);
-                    historyListOldHistory = em.merge(historyListOldHistory);
-                }
-            }
             for (History historyListNewHistory : historyListNew) {
                 if (!historyListOld.contains(historyListNewHistory)) {
-                    historyListNewHistory.getVmSettingList().add(vmSetting);
+                    VmSetting oldVmSettingIdOfHistoryListNewHistory = historyListNewHistory.getVmSettingId();
+                    historyListNewHistory.setVmSettingId(vmSetting);
                     historyListNewHistory = em.merge(historyListNewHistory);
+                    if (oldVmSettingIdOfHistoryListNewHistory != null && !oldVmSettingIdOfHistoryListNewHistory.equals(vmSetting)) {
+                        oldVmSettingIdOfHistoryListNewHistory.getHistoryList().remove(historyListNewHistory);
+                        oldVmSettingIdOfHistoryListNewHistory = em.merge(oldVmSettingIdOfHistoryListNewHistory);
+                    }
                 }
             }
             em.getTransaction().commit();
@@ -108,7 +125,7 @@ public class VmSettingJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -121,10 +138,16 @@ public class VmSettingJpaController implements Serializable {
             catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The vmSetting with id " + id + " no longer exists.", enfe);
             }
-            List<History> historyList = vmSetting.getHistoryList();
-            for (History historyListHistory : historyList) {
-                historyListHistory.getVmSettingList().remove(vmSetting);
-                historyListHistory = em.merge(historyListHistory);
+            List<String> illegalOrphanMessages = null;
+            List<History> historyListOrphanCheck = vmSetting.getHistoryList();
+            for (History historyListOrphanCheckHistory : historyListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This VmSetting (" + vmSetting + ") cannot be destroyed since the History " + historyListOrphanCheckHistory + " in its historyList field has a non-nullable vmSettingId field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(vmSetting);
             em.getTransaction().commit();
